FovScript
This Script creates a cone and looks for any objects within that cone.
Right now, I had to use an array like the tutorial I searched up was using, despite the fact the player character is the only one thats suppose to be there. This is just gonna be the initial version.