Weird Update (but there are some actual changes)
For some weird reason, the files that are on here aren't up to date. Oh well, I need to upload these files anyway since they have new stuff on them.
Biggest change is that fovScript now has some rough code to make a viewcone mesh appear in game! It's janky around the edges, which I may or may not fix depending on if its worth it or not, and some parts aren't filled...who knows?